feat: request error composer
