+++ packages/clients/src/internal/interceptors/composer.ts
@@ -1,3 +1,5 @@
+import type { RequestErrorInterceptor } from './types'
+
 export interface Interceptor<T> {
   (instance: Readonly<T>): T | Promise<T>
 }
@@ -19,3 +21,25 @@ export const composeInterceptors =
         resolve(instance)
       }),
     )
+
+/**
+ * Compose request error interceptors.
+ *
+ * @internal
+ */
+export const composeRequestErrorInterceptors =
+  (interceptors: RequestErrorInterceptor[]) =>
+  async (error: unknown): Promise<unknown> => {
+    let prevError = error
+    for (const interceptor of interceptors) {
+      try {
+        const res = await interceptor(prevError)
+
+        return res
+      } catch (err) {
+        prevError = err
+      }
+    }
+
+    throw error
+  }
