+++ packages/clients/src/internal/interceptors/helpers.ts
@@ -1,25 +1,4 @@
-export interface RequestInterceptor {
-  (request: Readonly<Request>): Request | Promise<Request>
-}
-
-export interface RequestErrorInterceptor {
-  (error: unknown): unknown | Promise<unknown>
-}
-
-export interface ResponseInterceptor {
-  (response: Readonly<Response>): Response | Promise<Response>
-}
-
-export interface ResponseErrorInterceptor {
-  (request: Request, error: unknown): unknown | Promise<unknown>
-}
-
-export type NetworkInterceptors = {
-  request?: RequestInterceptor
-  requestError?: RequestErrorInterceptor
-  response?: ResponseInterceptor
-  responseError?: ResponseErrorInterceptor
-}
+import type { RequestInterceptor } from './types'
 
 /**
  * Adds an header to a request through an interceptor.
